refactor: 브리핑 영상·연관 관광지 UI를 core:ui로 이동
