+++ mobile/core/ui/src/commonMain/kotlin/com/on/turip/core/ui/model/region/RelatedSpotModel.kt
@@ -1,15 +1,17 @@
-package com.on.turip.feature.randomtravel.impl.model
+package com.on.turip.core.ui.model.region
 
 import androidx.compose.runtime.Immutable
+import com.on.turip.core.model.region.RelatedSpotCategory
 import kotlinx.collections.immutable.ImmutableList
+import kotlinx.collections.immutable.toImmutableList
 
 /**
- * 브리핑의 연관 관광지 카드 한 장. 카테고리(관광지/숙박/음식 등) 하나가 카드 하나가 된다.
+ * 연관 관광지 카드 한 장. 카테고리(관광지/숙박/음식 등) 하나가 카드 하나가 된다.
  *
  * 첫 장소를 대표로 크게 보여주고 나머지는 이어붙여 노출한다.
  */
 @Immutable
-data class RandomTravelRelatedSpotModel(
+data class RelatedSpotModel(
     val category: String,
     val spots: ImmutableList<String>,
 ) {
@@ -23,3 +25,9 @@ data class RandomTravelRelatedSpotModel(
         private const val SPOT_SEPARATOR: String = " · "
     }
 }
+
+fun RelatedSpotCategory.toRelatedSpotModel(): RelatedSpotModel =
+    RelatedSpotModel(
+        category = category,
+        spots = spots.toImmutableList(),
+    )
