+++ mobile/feature/randomtravel/impl/src/commonMain/kotlin/com/on/turip/feature/randomtravel/impl/RandomTravelState.kt
@@ -3,10 +3,10 @@ package com.on.turip.feature.randomtravel.impl
 import androidx.compose.runtime.Immutable
 import com.on.turip.core.ui.UiState
 import com.on.turip.core.ui.error.ErrorUiState
+import com.on.turip.core.ui.model.content.VideoSummaryModel
 import com.on.turip.core.ui.model.namestatus.TuripNameStatusModel
+import com.on.turip.core.ui.model.region.RelatedSpotsUiState
 import com.on.turip.feature.randomtravel.impl.model.RandomDestinationModel
-import com.on.turip.feature.randomtravel.impl.model.RandomTravelRelatedSpotModel
-import com.on.turip.feature.randomtravel.impl.model.RandomTravelVideoModel
 import com.on.turip.feature.randomtravel.impl.model.TuripDraftPlaceModel
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
@@ -29,28 +29,6 @@ sealed interface RandomTravelPhase {
 
     /** 티켓 출력 + 여행 브리핑 */
     data object Briefing : RandomTravelPhase
-}
-
-/**
- * 연관 관광지 섹션의 상태.
- *
- * 서버가 일부 지역만 지원하고 조회도 실패할 수 있는데, 이때 섹션을 조용히 숨기면
- * `아직 구현되지 않은 것`과 구분할 수 없다. 그래서 사유를 상태로 구분해 항상 무언가를 그린다.
- */
-@Immutable
-sealed interface RelatedSpotsUiState {
-    data object Loading : RelatedSpotsUiState
-
-    data class Success(
-        val relatedSpots: ImmutableList<RandomTravelRelatedSpotModel>,
-    ) : RelatedSpotsUiState {
-        val isEmpty: Boolean = relatedSpots.isEmpty()
-    }
-
-    /** 서버가 지원하지 않는 지역 (400 `REGION_CATEGORY_INVALID`) */
-    data object Unsupported : RelatedSpotsUiState
-
-    data object Error : RelatedSpotsUiState
 }
 
 /**
@@ -90,7 +68,7 @@ data class RandomTravelState(
     val phase: RandomTravelPhase = RandomTravelPhase.Preparing,
     val reelNames: ImmutableList<String> = persistentListOf(),
     val destination: RandomDestinationModel? = null,
-    val videos: ImmutableList<RandomTravelVideoModel> = persistentListOf(),
+    val videos: ImmutableList<VideoSummaryModel> = persistentListOf(),
     val relatedSpotsUiState: RelatedSpotsUiState = RelatedSpotsUiState.Loading,
     val selectedContentId: Long? = null,
     val isBriefingLoading: Boolean = false,
