+++ mobile/feature/randomtravel/impl/src/commonMain/kotlin/com/on/turip/feature/randomtravel/impl/model/RandomTravelMapper.kt
@@ -1,25 +1,7 @@
 package com.on.turip.feature.randomtravel.impl.model
 
-import com.on.turip.core.model.content.video.VideoInformation
 import com.on.turip.core.model.region.RegionCategory
-import com.on.turip.core.model.region.RelatedSpotCategory
 import com.on.turip.core.model.trip.ContentPlace
-import com.on.turip.core.ui.util.TuripUrlConverter.convertVideoThumbnailUrl
-import kotlinx.collections.immutable.toImmutableList
-
-internal fun VideoInformation.toUiModel(): RandomTravelVideoModel =
-    RandomTravelVideoModel(
-        contentId = content.id,
-        title = content.videoData.title,
-        thumbnailUrl = convertVideoThumbnailUrl(content.videoData.url),
-        channelName = content.creator.channelName,
-        profileImageUrl = content.creator.profileImage,
-        uploadedDate = content.videoData.uploadedDate,
-        cityName = content.city.name,
-        nights = trip.tripDuration.nights,
-        days = trip.tripDuration.days,
-        placeCount = trip.tripPlaceCount,
-    )
 
 /**
  * 국내/해외는 [RegionCategory] 자체로는 알 수 없다.
@@ -34,12 +16,6 @@ internal fun RegionCategory.toUiModel(
         imageUrl = imageUrl,
         isDomestic = isDomestic,
         videoCount = videoCount,
-    )
-
-internal fun RelatedSpotCategory.toUiModel(): RandomTravelRelatedSpotModel =
-    RandomTravelRelatedSpotModel(
-        category = category,
-        spots = spots.toImmutableList(),
     )
 
 /**
